+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added support for Enclosure `:refresh` and `:reconfigure`
   - Added support to Ethernet network `:bulk_create`
   - Added support to Logical Enclosure `:reconfigure`, `:set_script`
+  - Added support to Network set actions
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0